Fix: Do setsockopt Correctly in Windows Including SSL_accept Timeouts
listener.cc
 - Set Windows socket timeouts correctly for SSL_accept:
     In Linux, "setsockopt" with SOL_SOCKET/SO_SNDTIMEO and
     SOL_SOCKET/SO_RCVTIMEO takes a pointer to a "struct timeval" as
     the parameter. In Windows however the parameter is a pointer to a
     DWORD value, the value being the number of milliseconds of the
     timeout. Setting these socket timeouts is needed before and after
     doing SSL_accept. Calling setsockopt to set the timeouts, but
     with the wrong values, was making the SSL_accept call effectively
     non-blocking in the Windows case, causing an occasional error
     from SSL_accept with ssl-error SSL_ERROR_SYSCALL indicating that
     the SSL/TLS handshake was incomplete at the moment of the
     SSL_accept call and that SSL_accept was not then waiting for the
     handshake to complete before returning.
 - For Windows, use pist_sock_set_timeout when setting socket timeouts
 - Check return code for make_non_blocking even in non-debug build
 - Don't set CLOEXEC socket flag in Windows case
 - Use "unsigned long int" for millisecond count; also in pist_sockfns
 - Add additional debug logging
 - Remove some unneeded blank lines

winornix.h
 - Use correct type for get/setsockopt optval in Windows
     In Linux, the 4th parameter of setsockopt, a pointer to optval,
     is a most commonly an "int *", and the most common type for the
     optval is an "int". However in Windows, the 4th parameter of
     setsockopt, a pointer to optval, is a "char *", and the most
     common type for the optval is an "DWORD" (i.e. uint32_t).
     Previously, we were using the Linux types (notably "int" for
     optval) even in Windows. The changes here make us use the Windows
     types (notably DWORD) in Windows, while continuing to use the
     Linux types in Linux.
 - Define PST_SOCK_OPT_VAL_PTR_T (type for 4th parm of setsockopt)
 - Define PST_SOCK_OPT_VAL_TYPICAL_T (the most common type of optval)
 - Use type-alias keyword "using", not typedef, for get/setsockopt

pist_sockfns.cc
transport.cc
listener.cc
listener_test.cc
 - Use PST_SOCK_OPT_VAL_PTR_T and PST_SOCK_OPT_VAL_TYPICAL_T

pist_sockfns.h/.cc
 - Add a function pist_sock_set_timeout for Windows (used by listener)

listener_test.cc
 - Ran clang-format (apparently not run previously)

listener_tls_test.cc
 - Adjust the listener_tls_test.cc/tls_handshake_timeout_custom test,
   which tests for how long a timeout takes, to adjust for the
   apparent behavior in Windows where BIO_do_connect may itself take
   some non-trivial period of time; previously was causing an
   occasional spurious test failure in Windows.

.clang-format-ignore
 - Ignore pist_sockfns.cc

commonsetup.ps1
 - Use -UseBasicParsing option to get doxygen release: Without that
   option, Invoke-WebRequest hangs in Windows Server 2019 (PS
   v5.1.17763.2931/Desktop) for that URI, apparently a known
   PowerShell bug.

+++ include/pistache/pist_sockfns.h
@@ -61,7 +61,7 @@ int pist_sock_bind(em_socket_t em_sock, const struct sockaddr *addr,
 // On success, returns 0. On failure, -1 is returned and errno is set.
 int pist_sock_set_timeout(em_socket_t em_sock,
                           int optname, // SO_RCVTIMEO or SO_SNDTIMEO
-                          unsigned int timeout_in_ms);
+                          unsigned long int timeout_in_ms);
 
 // On success returns an em_socket_t for the accepted socket. On failure, -1 is
 // returned and errno is set.
